# Small perf adjustments for `ShortWeierstrassProjectivePoint` operations