--- a/math/src/elliptic_curve/short_weierstrass/point.rs
+++ b/math/src/elliptic_curve/short_weierstrass/point.rs
@@ -53,6 +53,9 @@ impl<E: IsShortWeierstrass> ShortWeierstrassProjectivePoint<E> {
     }
 
     pub fn double(&self) -> Self {
+        if self.is_neutral_element() {
+            return self.clone();
+        }
         let [px, py, pz] = self.coordinates();
 
         let px_square = px * px;
@@ -87,17 +90,17 @@ impl<E: IsShortWeierstrass> ShortWeierstrassProjectivePoint<E> {
     }
 
     pub fn operate_with_affine(&self, other: &Self) -> Self {
-        let [px, py, pz] = self.coordinates();
-        let [qx, qy, _qz] = other.coordinates();
-        let u = qy * pz;
-        let v = qx * pz;
-
         if self.is_neutral_element() {
             return other.clone();
         }
         if other.is_neutral_element() {
             return self.clone();
         }
+
+        let [px, py, pz] = self.coordinates();
+        let [qx, qy, _qz] = other.coordinates();
+        let u = qy * pz;
+        let v = qx * pz;
 
         if u == *py {
             if v != *px || *py == FieldElement::zero() {