# Small perf adjustments for `ShortWeierstrassProjectivePoint` operations

## Description

- In the `double` operation, the pull request implements an early return strategy to check if `self` is the neutral element before performing any further calculations. If `self` is indeed the neutral element, the function immediately returns a clone of `self`. This optimization enhances performance by avoiding unnecessary computations, resulting in faster execution.

- The pull request optimizes the `operate_with_affine` function by reordering the sequence of operations. Previously, the function performed certain calculations before checking if `self` or `other` were neutral elements. This resulted in unnecessary computation overhead. The pull request rectifies this by placing the neutral element check at the beginning of the function. If either `self` or `other` is the neutral element, the function terminates early, saving computational costs and improving overall efficiency.

## Type of change

Please delete options that are not relevant.

- [X] Optimization

## Checklist
- [ ] Linked to Github Issue
- [ ] Unit tests added
- [ ] This change requires new documentation.
  - [ ] Documentation has been added/updated.
- [X] This change is an Optimization
  - [ ] Benchmarks added/run