@tcoratger 2024-03-12
> I don't think the first optimization is right. The early return adds an additional branch, so it may make the average case slower, and we usually don't care about optimizing the neutral element case. You need to bench this and compare it, but I highly doubt the additional branch is worth.

Here is the benchmark for the double case (for the general situation):
- Before:
```shell
point Projective Double | Lambdaworks
                        time:   [364.85 ns 364.97 ns 365.09 ns]
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
```

- After:
```shell
point Projective Double | Lambdaworks
                        time:   [364.55 ns 365.17 ns 366.33 ns]
                        change: [-0.1313% +0.0010% +0.1993%] (p = 1.00 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) high mild
  6 (6.00%) high severe
```